Create global variable with MetricsLogger
Must be truly global (i.e. not just static) to avoid issue where tools and internals of Cognite3DViewer would access two different MetricsLogger instances
