+++ viewer/core/src/public/migration/Cognite3DModel.test.ts
@@ -8,14 +8,14 @@ import { NodesLocalClient } from '@reveal/nodes-api';
 import { DefaultNodeAppearance, TreeIndexNodeCollection } from '@reveal/cad-styling';
 import { CadMaterialManager, CadNode } from '@reveal/rendering';
 import { CadModelMetadata } from '@reveal/cad-parsers';
-import { initMetrics } from '@reveal/metrics';
+import { MetricsLogger } from '@reveal/metrics';
 import { createCadModelMetadata, generateSectorTree } from '../../../../test-utilities';
 
 describe(Cognite3DModel.name, () => {
   let model: Cognite3DModel;
 
   beforeAll(() => {
-    initMetrics(false, '', '', {});
+    MetricsLogger.initMetrics(false, '', '', {});
   });
 
   beforeEach(() => {
