+++ viewer/core/src/public/migration/Cognite3DViewer.ts
@@ -13,7 +13,7 @@ import { LoadingState } from '@reveal/cad-geometry-loaders';
 import { defaultRenderOptions, SsaoParameters, SsaoSampleQuality, AntiAliasingMode } from '@reveal/rendering';
 
 import { assertNever, clickOrTouchEventOffset, EventTrigger } from '@reveal/utilities';
-import { trackError, trackEvent } from '@reveal/metrics';
+import { MetricsLogger } from '@reveal/metrics';
 
 import { worldToNormalizedViewportCoordinates, worldToViewportCoordinates } from '../../utilities/worldToViewport';
 import { intersectCadNodes } from '../../datamodels/cad/picking';
@@ -292,7 +292,7 @@ export class Cognite3DViewer {
           }
         },
         error =>
-          trackError(error, {
+          MetricsLogger.trackError(error, {
             moduleName: 'Cognite3DViewer',
             methodName: 'constructor'
           })
@@ -301,7 +301,7 @@ export class Cognite3DViewer {
 
     this.animate(0);
 
-    trackEvent('construct3dViewer', {
+    MetricsLogger.trackEvent('construct3dViewer', {
       constructorOptions: omit(options, [
         'sdk',
         'domElement',
@@ -1592,6 +1592,8 @@ function createRevealManagerOptions(viewerOptions: Cognite3DViewerOptions): Reve
   const edgeDetectionParameters = {
     enabled: viewerOptions.enableEdges ?? defaultRenderOptions.edgeDetectionParameters.enabled
   };
+
+  revealOptions.logMetrics = viewerOptions.logMetrics;
 
   revealOptions.renderOptions = {
     antiAliasing,
