+++ viewer/packages/metrics/index.ts
@@ -1,14 +1,6 @@
 /*!
  * Copyright 2021 Cognite AS
  */
-export {
-  trackError,
-  trackLoadModel,
-  trackCameraNavigation,
-  trackEvent,
-  trackCreateTool,
-  trackCadModelStyled,
-  initMetrics,
-  EventProps,
-  TrackedEvents
-} from './src/metrics';
+export { MetricsLogger } from './src/MetricsLogger';
+
+export { EventProps, TrackedEvents } from './src/types';
