+++ viewer/packages/tools/src/ExplodedViewTool.ts
@@ -4,7 +4,7 @@
 import * as THREE from 'three';
 import { Cognite3DModel } from '@reveal/core';
 import { Cognite3DViewerToolBase } from './Cognite3DViewerToolBase';
-import { trackCreateTool } from '@reveal/metrics';
+import { MetricsLogger } from '@reveal/metrics';
 
 export class ExplodedViewTool extends Cognite3DViewerToolBase {
   private readonly _cadModel: Cognite3DModel;
@@ -23,7 +23,7 @@ export class ExplodedViewTool extends Cognite3DViewerToolBase {
 
     this.preloadBoundingBoxData(cadModel, treeIndex);
 
-    trackCreateTool('ExplodedViewTool');
+    MetricsLogger.trackCreateTool('ExplodedViewTool');
   }
 
   public async expand(expandRadius: number): Promise<void> {
