+++ viewer/packages/tools/src/Geomap/GeomapTool.ts
@@ -7,7 +7,7 @@ import { LatLongPosition, MapConfig } from './MapConfig';
 import { Geomap } from './Geomap';
 
 import { Cognite3DViewer } from '@reveal/core';
-import { trackCreateTool } from '@reveal/metrics';
+import { MetricsLogger } from '@reveal/metrics';
 
 /**
  * The `GeomapTool` is a geolocation for the models and allow the user to place them on the maps.
@@ -23,7 +23,7 @@ export class GeomapTool extends Cognite3DViewerToolBase {
     this._viewer = viewer;
     this._maps = new Geomap(this._viewer, config);
 
-    trackCreateTool('AxisViewTool');
+    MetricsLogger.trackCreateTool('AxisViewTool');
   }
 
   /**
