+++ viewer/packages/utilities/src/globals.ts
@@ -1,0 +1,5 @@
+/*!
+ * Copyright 2021 Cognite AS
+ */
+
+declare let revealMetricsLogger: { metricsLogger: any };
